three-tier-arch/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
psycopg2-binary==2.9.9
pymilvus==2.3.3
minio==7.2.0
sentence-transformers==2.2.2
python-multipart==0.0.6
